floating point accuracy error fixed

+++ sdl/src/AABB.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-bool AABB(int x, int y, int w, int h, int x2, int y2, int w2, int h2);
+bool AABB(float x, float y, float w, float h, float x2, float y2, float w2, float h2);
